+++ crates/matrix-sdk/CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- Add a new `RequestConfig::skip_auth()` option. This is useful to ensure that
+  certain request won't ever include an authorization header.
+  ([#5822](https://github.com/matrix-org/matrix-rust-sdk/pull/5822))
 - Add support for extended profile fields with `Account::fetch_profile_field_of()`,
   `Account::fetch_profile_field_of_static()`, `Account::set_profile_field()` and
   `Account::delete_profile_field()`.